extract shared assembly utilities
