+++ src/main/scala/org/hammerlab/guacamole/assembly/DeBruijnGraph.scala
@@ -1,5 +1,7 @@
 package org.hammerlab.guacamole.assembly
 
+import htsjdk.samtools.CigarOperator
+import org.hammerlab.guacamole.reads.MappedRead
 import org.hammerlab.guacamole.util.Bases
 
 import scala.collection.mutable
@@ -123,13 +125,13 @@ class DeBruijnGraph(val kmerSize: Int,
   }
 
   /**
-   *  Searches forward or backward from a node to find those connected by a unique path
+    *  Searches forward or backward from a node to find those connected by a unique path
    *
    * @param kmer Kmer to search from
-   * @param searchForward If true, search children of node, otherwise parents
-   * @param avoidLoops If true, only explore a node once
-   * @return List of kmers, that can be merged starting with `kmer`
-   */
+    * @param searchForward If true, search children of node, otherwise parents
+    * @param avoidLoops If true, only explore a node once
+    * @return List of kmers, that can be merged starting with `kmer`
+    */
   private[assembly] def findMergeable(kmer: Kmer,
                                       searchForward: Boolean,
                                       avoidLoops: Boolean = true): Seq[Kmer] = {
@@ -167,16 +169,16 @@ class DeBruijnGraph(val kmerSize: Int,
   type Path = List[Kmer]
 
   /**
-   * Find a path from source to sink in the graph
+    * Find a path from source to sink in the graph
    *
    * @param source Kmer node to begin search
-   * @param sink Kmer to search for
-   * @param minPathLength Minimum number of kmers to traverse before finding the sink
-   * @param maxPathLength Maximum number of kmers to traverse before finding the sink
-   * @param maxPaths Maximum number of paths to find from source to sink
-   * @param avoidLoops If avoiding loops, skip nodes that have already been visited
-   * @return Set
-   */
+    * @param sink Kmer to search for
+    * @param minPathLength Minimum number of kmers to traverse before finding the sink
+    * @param maxPathLength Maximum number of kmers to traverse before finding the sink
+    * @param maxPaths Maximum number of paths to find from source to sink
+    * @param avoidLoops If avoiding loops, skip nodes that have already been visited
+    * @return Set
+    */
   def depthFirstSearch(source: Kmer,
                        sink: Kmer,
                        minPathLength: Int = 1,
@@ -301,26 +303,35 @@ class DeBruijnGraph(val kmerSize: Int,
    *
    * @return
    */
-  def roots: Iterable[Kmer] = {
-    kmerCounts.keys.filter(parents(_).isEmpty)
+  def sources: Iterable[Kmer] = {
+    kmerCounts.keys.filter(parents(_).isEmpty).map(_.take(kmerSize))
   }
 
   /**
-   * Find all children of a given node
+   * Find all nodes that have out-degree = 0
+   *
+   * @return
+   */
+  def sinks: Iterable[Kmer] = {
+    kmerCounts.keys.filter(children(_).isEmpty).map(_.takeRight(kmerSize))
+  }
+
+  /**
+    * Find all children of a given node
    *
    * @param node Kmer to find parents of
-   * @return  List of Kmers where their prefix matches this node's suffix
-   */
+    * @return  List of Kmers where their prefix matches this node's suffix
+    */
   def children(node: Kmer): List[Kmer] = {
     prefixTable.getOrElse(kmerSuffix(node), List.empty)
   }
 
   /**
-   * Find all parents of a given node
+    * Find all parents of a given node
    *
    * @param node Kmer to find parents of
-   * @return List of Kmers where their suffix matches this nodes prefix
-   */
+    * @return List of Kmers where their suffix matches this nodes prefix
+    */
   def parents(node: Kmer): List[Kmer] = {
     suffixTable.getOrElse(kmerPrefix(node), List.empty)
   }
@@ -330,6 +341,7 @@ class DeBruijnGraph(val kmerSize: Int,
 object DeBruijnGraph {
   type Sequence = DeBruijnGraph#Sequence
   type Kmer = DeBruijnGraph#Kmer
+
   def apply(sequences: Seq[Sequence],
             kmerSize: Int,
             minOccurrence: Int = 1,
@@ -337,7 +349,8 @@ object DeBruijnGraph {
 
     val kmerCounts = mutable.Map.empty[DeBruijnGraph#Kmer, Int]
 
-    sequences.filter(Bases.allStandardBases(_))
+    sequences
+      .filter(Bases.allStandardBases(_))
       .foreach(
         _.sliding(kmerSize)
           .foreach(seq => {
@@ -354,17 +367,125 @@ object DeBruijnGraph {
   }
 
   /**
-   * Merge sequences where we expect consecutive entries to overlap by `overlapSize` bases.
-   *
-   * @param sequences Set of sequences that we would like to combine into a single sequence
-   * @param overlapSize The amount of the sequences we expect to overlap (The number of bases the last sequence overlaps
-   *                    with the next. For a standard kmer graph, this is the length of the kmer length - 1
-   * @return A single merged sequence
-   */
+    * Merge sequences where we expect consecutive entries to overlap by `overlapSize` bases.
+    *
+    * @param sequences Set of sequences that we would like to combine into a single sequence
+    * @param overlapSize The amount of the sequences we expect to overlap (The number of bases the last sequence overlaps
+    *                    with the next. For a standard kmer graph, this is the length of the kmer length - 1
+    *
+    * @return A single merged sequence
+    */
   def mergeOverlappingSequences(sequences: Seq[Sequence], overlapSize: Int): Sequence = {
     val head = sequences.headOption.getOrElse(Seq.empty)
     val rest = sequences.tail.flatMap(sequence => sequence.takeRight(sequence.length - overlapSize + 1))
     head ++ rest
+  }
+
+  /**
+   * Extract a subsequence from a read sequence
+   *
+   * NOTE: This assumes that the read entirely covers the reference region and does not check for insertions
+   * or deletion
+   *
+   * @param read Read to extract subsequence from
+   * @param startLocus Start (inclusive) locus on the reference
+   * @param endLocus End (exclusive) locus on the reference
+   * @return Subsequence overlapping [startLocus, endLocus)
+   */
+  private def getSequenceFromRead(read: MappedRead, startLocus: Int, endLocus: Int) = {
+    read.sequence.slice(startLocus - read.unclippedStart.toInt, endLocus - read.unclippedStart.toInt)
+  }
+
+  /**
+   * For a given set of reads identify all kmers that appear in the specified reference region
+   *
+   * @param reads  Set of reads to extract sequence from
+   * @param startLocus Start (inclusive) locus on the reference
+   * @param endLocus End (exclusive) locus on the reference
+   * @param minOccurrence Minimum number of times a subsequence needs to appear to be included
+   * @return List of subsequences overlapping [startLocus, endLocus) that appear at least `minOccurrence` time
+   */
+  private def getConsensusKmer(reads: Seq[MappedRead],
+                               startLocus: Int,
+                               endLocus: Int,
+                               minOccurrence: Int): Iterable[Vector[Byte]] = {
+
+    // Filter to reads that entirely cover the region
+    // Exclude reads that have any non-M Cigars (these don't have a 1 to 1 base mapping to the region)
+    val overlapping = reads
+      .filter(!_.cigarElements.exists(_.getOperator != CigarOperator.M))
+      .filter(r => r.overlapsLocus(startLocus) && r.overlapsLocus(endLocus - 1))
+
+    // Extract the sequences from each region
+    val sequences = overlapping
+      .map(r => getSequenceFromRead(r, startLocus, endLocus))
+
+    // Filter to sequences that appear at least `minOccurrence` times
+    sequences
+      .groupBy(identity)
+      .map(kv => (kv._1, kv._2.length))
+      .filter(_._2 >= minOccurrence)
+      .map(_._1.toVector)
+  }
+
+  /**
+   * Find paths through the reads given that represent the sequence covering referenceStart and referenceEnd
+   *
+   * @param reads Reads to use to build the graph
+   * @param referenceStart Start of the reference region corresponding to the reads
+   * @param referenceEnd End of the reference region corresponding to the reads
+   * @param referenceSequence Reference sequence overlapping [referenceStart, referenceEnd)
+   * @param kmerSize Length of kmers to use to traverse the paths
+   * @param minOccurrence Minimum number of occurrences of the each kmer
+   * @param maxPaths Maximum number of paths to find
+   * @param debugPrint Print debug statements (default: false)
+   * @return List of paths that traverse the region
+   */
+  def discoverPathsFromReads(reads: Seq[MappedRead],
+                             referenceStart: Int,
+                             referenceEnd: Int,
+                             referenceSequence: Array[Byte],
+                             kmerSize: Int,
+                             minOccurrence: Int,
+                             maxPaths: Int,
+                             debugPrint: Boolean = false) = {
+    val referenceKmerSource = referenceSequence.take(kmerSize).toVector
+    val referenceKmerSink = referenceSequence.takeRight(kmerSize).toVector
+
+    val currentGraph: DeBruijnGraph = DeBruijnGraph(
+      reads.map(_.sequence),
+      kmerSize,
+      minOccurrence,
+      mergeNodes = true
+    )
+
+    val sources: Set[Vector[Byte]] = (getConsensusKmer(
+      reads,
+      referenceStart,
+      referenceStart + kmerSize,
+      minOccurrence = minOccurrence
+    ) ++ Seq(referenceKmerSource)).toSet
+
+    val sinks: Set[Vector[Byte]] = (getConsensusKmer(
+      reads,
+      referenceEnd - kmerSize,
+      referenceEnd,
+      minOccurrence = minOccurrence
+    ) ++ Seq(referenceKmerSink)).toSet
+
+    for {
+      source <- sources
+      sink <- sinks
+      if source != sink
+      path <- currentGraph.depthFirstSearch(
+        source,
+        sink,
+        maxPaths = maxPaths,
+        debugPrint = debugPrint
+      )
+    } yield {
+      path
+    }
   }
 
 }
